Moved code.pcf to Harness project
code.pcf was originally placed in the Pint project because it's loaded
by Pint and it needed to be in Pint's directory. But since Pint is a
library, it inherits the directory of whatever program includes it (in
this case, Harness), so code.pcf needs to be placed in Harness'
directory instead.